refactor(lib:design-system): :recycle: #1510 - update indicator: error message
