--- libs/design-system/error-indicator/src/lib/error-indicator.component.html
@@ -1,6 +1,0 @@
-<mat-icon>error</mat-icon>
-<div class="errors">
-  @for (error of errors(); track error) {
-    <div class="error">{{ error }}</div>
-  }
-</div>
