+++ libs/design-system/indicators/error-indicator/ng-package.json
