+++ libs/design-system/indicators/error-indicator/src/index.ts
