+++ libs/design-system/indicators/error-indicator/src/lib/error-indicator.component.html
@@ -1,0 +1,12 @@
+<mat-icon>error</mat-icon>
+<div class="errors">
+  @for (error of errors(); track error) {
+    <div class="error">
+      {{ error }}
+
+      @if (actionLink()) {
+        <a class="link" hraHyperlink [attr.href]="actionLink()">{{ actionLinkLabel() }}</a>
+      }
+    </div>
+  }
+</div>
