+++ libs/design-system/indicators/error-indicator/src/lib/error-indicator.component.spec.ts
