+++ libs/design-system/indicators/error-indicator/src/lib/error-indicator.component.stories.ts
@@ -1,10 +1,10 @@
 import { applicationConfig, Meta, StoryObj } from '@storybook/angular';
-import { provideDesignSystem } from '../../../src';
+import { provideDesignSystem } from '../../../../src';
 import { ErrorIndicatorComponent } from './error-indicator.component';
 
 const meta: Meta = {
   component: ErrorIndicatorComponent,
-  title: 'ErrorIndicatorComponent',
+  title: 'Design System/Indicators/Error Indicator',
   parameters: {
     design: {
       type: 'figma',
@@ -16,6 +16,20 @@ const meta: Meta = {
       providers: [provideDesignSystem()],
     }),
   ],
+  argTypes: {
+    errors: {
+      control: 'object',
+      description: 'List of errors to be shown in the indicator',
+    },
+    actionLink: {
+      control: 'text',
+      description: 'Call to action link',
+    },
+    actionLinkLabel: {
+      control: 'text',
+      description: 'Call to action link label',
+    },
+  },
 };
 export default meta;
 type Story = StoryObj;
@@ -29,5 +43,13 @@ export const SingleError: Story = {
 export const MultipleErrors: Story = {
   args: {
     errors: ['Required columns missing: Column Name, Column Name', 'Please upload a file with all required columns.'],
+  },
+};
+
+export const WithLink: Story = {
+  args: {
+    errors: ['Unexpected Error Occurred. Please contact support.'],
+    actionLink: 'https://example.com',
+    actionLinkLabel: 'Submit a ticket',
   },
 };
