+++ libs/design-system/styles/_light-theme.scss
@@ -102,14 +102,14 @@ $_palettes: (
     10: #2f1400,
     20: #690005,
     25: #7e0007,
-    30: #93000a,
+    30: #2f1400,
     35: #a80710,
     40: #883700,
     50: #de3730,
     60: #ff5449,
     70: #ff897d,
     80: #ffb4ab,
-    90: #ffdcc4,
+    90: #ffe8d8,
     95: #ffedea,
     98: #fff8f7,
     99: #fffbff,
